third_party/blink/web_tests/external/wpt/css/css-shapes/shape-outside/values/shape-outside-circle-010-expected.txt

This is a testharness.js-based test.
FAIL circle(calc(10in) at 50% 50%) - inline style assert_equals: expected "circle(calc(960px) at 50% 50%)" but got "circle(calc(10in) at 50% 50%)"
PASS circle(calc(10in + 20px) at 50% 50%) - inline style
PASS circle(calc(30%) at 50% 50%) - inline style
PASS circle(calc(100%/4) at 50% 50%) - inline style
PASS circle(calc(25%*3) at 50% 50%) - inline style
FAIL circle(calc(25%*3 - 10in) at 50% 50%) - inline style assert_equals: expected "circle(calc(75% - 960px) at 50% 50%)" but got "circle(calc(75% - 10in) at 50% 50%)"
FAIL circle(calc((12.5%*6 + 10in) / 4) at 50% 50%) - inline style assert_equals: expected "circle(calc(18.75% + 240px) at 50% 50%)" but got "circle(calc((75% + 10in) / 4) at 50% 50%)"
PASS circle(calc(10in) at 50% 50%) - computed style
PASS circle(calc(10in + 20px) at 50% 50%) - computed style
PASS circle(calc(30%) at 50% 50%) - computed style
PASS circle(calc(100%/4) at 50% 50%) - computed style
PASS circle(calc(25%*3) at 50% 50%) - computed style
PASS circle(calc(25%*3 - 10in) at 50% 50%) - computed style
PASS circle(calc((12.5%*6 + 10in) / 4) at 50% 50%) - computed style
Harness: the test ran to completion.
